webserver/requirements.txt
Django==2.1.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.3
elasticsearch==7.0.0
elasticsearch-dsl==7.0.0

numpy==1.16.3
Pillow==6.0.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
tensorflow==1.13.1
opencv-python

requests==2.20.1
urllib3==1.22


chardet
matplotlib
